Integrate Inputs like Spotify
Problem is that spotify etc. is not an input, it is a "feature" in AVR

The core problem is, that the yamaha-nodejs has no functionality to read out the features.
So I started a pull request to yamaha-nodejs.
In the meantime I added the functionality in the "simpleCommands_addition.js".